Fix cache invalidation
Cache invalidation breaks when querying the same feature with different
select params.

Thus, select needs to be part of the feature cache key, which makes
cache invalidation a lot harder. So use the same technique as query
cache to invalidate feature caches.

Also sneak in a hack to bypass the cache